Update jquery_checklist.js
Resolve a conflict when using both the Checklist tweak and the Icon Replacement tweak.
Changes the target to find an item with 'Checklist' in the name, but ignore it if it also contains the word 'Replacement'.  This way it will not find the 'Replacement Icon: Checklist' tweak item instead of the actual 'Checklist' item.